Refactored code and fixed endpoints; need to modify tests
