+++ instadam/project.py
@@ -1,4 +1,3 @@
-
 from flask import Blueprint, jsonify, request, abort
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from sqlalchemy.exc import IntegrityError, DatabaseError
@@ -6,6 +5,7 @@ from sqlalchemy.exc import IntegrityError, DatabaseError
 from flask import current_app as app
 from instadam.models.project_permission import AccessTypeEnum, ProjectPermission
 from instadam.models.user import User, PrivilegesEnum
+from instadam.models.image import Image
 from .app import db
 from .models.project import Project
 import os
@@ -55,14 +55,16 @@ def create_project():
         db.session.flush()
     except IntegrityError:
         db.session.rollback()
-        abort(400, 'Duplicate project name. '
-                   'Please provide a different project name.')
+        abort(
+            400, 'Duplicate project name. '
+            'Please provide a different project name.')
     else:
         # if able to add project to db, try add project_permission
         # creator is granted with READ_WRITE privilege
-        project_permission = ProjectPermission(access_type=AccessTypeEnum.READ_WRITE,
-                                               user_id=user_id,
-                                               project_id=project.id)
+        project_permission = ProjectPermission(
+            access_type=AccessTypeEnum.READ_WRITE,
+            user_id=user_id,
+            project_id=project.id)
         try:
             db.session.add(project_permission)
             db.session.flush()
@@ -73,8 +75,61 @@ def create_project():
     db.session.commit()
 
     # create project directory
-    project_dir = os.path.join(app.config['STATIC_STORAGE_DIR'], str(project.id))
+    project_dir = os.path.join(app.config['STATIC_STORAGE_DIR'],
+                               str(project.id))
     os.makedirs(project_dir)
 
-    return jsonify({'msg': 'project added successfully',
-                    'project_id': project.id}), 201
+    return jsonify({
+        'msg': 'project added successfully',
+        'project_id': project.id
+    }), 201
+
+
+@bp.route('/projects/<project_id>/unannotated', methods=['GET'])
+@jwt_required
+def get_unannotated_images(project_id):
+    """
+    Get unannotated images across ALL projects so that user (annotator) can
+    see images to annotate
+    NOTE: Only returning a fixed number of images (k=5) for Iteration 3
+    """
+    unannotated_images = Image.query.filter_by(is_annotated=False).all()[0:k]
+    if not unannotated_images:
+        return jsonify({'unannotated_images': []}), 200
+
+    unannotated_images_res = []
+    for unannotated_image in unannotated_images:
+        unannotated_image_res = {}
+        unannotated_image_res['id'] = unannotated_image.id
+        unannotated_image_res['name'] = unannotated_image.image_name
+        unannotated_image_res['path'] = unannotated_image.image_path
+        unannotated_image_res['project_id'] = unannotated_image.project_id
+        unannotated_images_res.append(unannotated_image_res)
+
+    return jsonify({'unannotated_images': unannotated_images_res}), 200
+
+
+@bp.route('/projects/<project_id>/images')
+@jwt_required
+def get_project_images(project_id):
+    """
+    Get all images (annotated and unannotated) of project with project_id
+    NOTE: Only returning a fixed number of images (k=5) for Iteration 3
+
+    Args:
+        project_id: The id of the project
+    """
+    project_images = Image.query.filter_by(project_id=project_id).all()[0:k]
+    if not project_images:
+        return jsonify({'project_images': []}), 200
+
+    project_images_res = []
+    for project_image in project_images:
+        project_image_res = {}
+        project_image_res['id'] = project_image.id
+        project_image_res['name'] = project_image.image_name
+        project_image_res['path'] = project_image.image_path
+        project_image_res['project_id'] = project_image.project_id
+        project_images_res.append(project_image_res)
+
+    return jsonify({'project_images': project_images_res}), 200
